Store data using Postgres ARRAY (#90)
Data are now stored in Postgres ARRAYs. 

Searching for data should be much quicker (especially for large numbers of large datasets).

Segmenting data done in Python - we may run into memory limits here when asked for contents of large columns.

**Data now retrieved as a newline-separated stream from `/columns/:id/values/`**

Updates to the frontend to handle fetching data as a stream.